bridging-plugin: refactor
