+++ CMakeLists.txt
@@ -15,19 +15,24 @@ include(CompileOptions.cmake)
 
 set(PLUGIN 0 CACHE BOOL "Build a plugin")
 
-include_directories(${CMAKE_SOURCE_DIR}/src/ ${CMAKE_SOURCE_DIR}/deps/uthash/include/)
-
 set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} "${CMAKE_SOURCE_DIR}/CMakeModules")
 find_package(SYSREPO REQUIRED)
 find_package(LIBYANG REQUIRED)
 find_package(SRPC REQUIRED)
 find_package(NL REQUIRED)
 
-include_directories(
-    ${SYSREPO_INCLUDE_DIRS}
-    ${LIBYANG_INCLUDE_DIRS}
-    ${SRPC_INCLUDE_DIRS}
-    ${NL_INCLUDE_DIRS})
+# get sysrepo version
+find_package(PkgConfig)
+if(PKG_CONFIG_FOUND)
+    execute_process(COMMAND ${PKG_CONFIG_EXECUTABLE} "--modversion" "sysrepo" OUTPUT_VARIABLE SYSREPO_VERSION)
+    if(SYSREPO_VERSION)
+        # strip new line from string
+        string(STRIP ${SYSREPO_VERSION} SYSREPO_VERSION)
+        if(${SYSREPO_VERSION} VERSION_LESS "1.0.0")
+            message(FATAL_ERROR "${PROJECT_NAME} requires at least libsysrepo verision 1.0.0")
+        endif()
+    endif()
+endif()
 
 option(INTERFACES_PLUGIN "Enable interfaces plugin" OFF)
 option(ROUTING_PLUGIN "Enable routing plugin" OFF)
