+++ src/bridging/plugin/data/component.h
@@ -1,0 +1,9 @@
+#ifndef BRIDGING_PLUGIN_DATA_COMPONENT_H
+#define BRIDGING_PLUGIN_DATA_COMPONENT_H
+
+#include "plugin/types.h"
+
+void bridge_component_init(bridge_component_t *c);
+void bridge_component_free(bridge_component_t *c);
+
+#endif // BRIDGING_PLUGIN_DATA_COMPONENT_H
